# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 1Vault Backend API
  version: 1.2.2
  description: |
    REST API for 1Vault (Solana vaults). Pass `cluster=devnet|mainnet-beta` or header `X-1Vault-Cluster`.

    **Frontend integration guide:** `backend/docs/FRONTEND_GUIDE.md` (vault types, 1vault Licence / 1VL,
    withdraw account layout, flow modes, simulator-v2 canvas mapping).

    **Groups**
    - **Token · Market** — spot price / kline chart
    - **Token · Research** — token info / security / pool / holders / traders / holder analysis
    - **Wallet · Analytics** — portfolio, activity, stats, created tokens, copy-trade score
    - **Discover** — trending profiles, boosts, metas, pair search
    - **Live · Stream** — WebSocket live feeds (profiles / boosts / takeovers / ads)
    - **Account · Wallets** — bind Solana wallet to logged-in user
    - **Tx · *** — unsigned tx prep (client signs → `/v1/tx/submit`)
    - **Flows** — multi-step async jobs

    Token market, research, and wallet analytics require market data to be configured.
    Token metadata (`name` / `symbol` / links) is untrusted user-facing data.

    **Wallet kind:** `walletKind=eoa` (individual/personal) vs `walletKind=pda` (vault wallet).
    See Wallet · Analytics tag and `GET /v1/wallets/{walletAddress}/kind`.

    **Roles:** `strategies` (formerly degen) and `investors` (formerly retail). Legacy `degen`/`retail` still accepted on write.

...
servers:
  - url: http://localhost:3090
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    WalletAddressPath:
      name: walletAddress
      in: path
      required: true
      schema: { type: string }
      description: Solana wallet address (EOA individual or vault PDA)
...
paths:
...
  /v1/protocol:
    get:
      tags: [Protocol]
      summary: Protocol addresses & DEX program IDs
      description: |
        Returns on-chain program IDs, PDAs, licence mint, and configured DEX programs for the selected cluster.

        **Licence (1VL):** `licenseMint`, `licenseLockAmount` (per-vault lock on create_vault).
        Display as **1vault Licence** — default lock 1,000,000 tokens. See `LicenseFields` schema and FRONTEND_GUIDE.md §3.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
      responses:
        "200":
          description: Protocol addresses
  /v1/protocol/state:
    get:
      tags: [Protocol]
      summary: Protocol on-chain / DB state snapshot
      description: |
        Snapshot of protocol-level state from DB/indexer for the selected cluster. Useful for dashboards and ops checks.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
      responses:
        "200":
          description: Protocol state
  /v1/tokens/{mint}/price:
    get:
      tags: [Token · Market]
      summary: Get real-time USD spot price
      description: |
        Returns the current USD price for a token mint (Solana mainnet market data). Optionally pass `amount` + `exitBps` to estimate sell notional in USD and SOL proceeds (lamports) — used to auto-fill exit/reduce proceeds.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: mint
          in: path
          required: true
          schema: { type: string }
          description: Token mint address (Solana)
        - name: amount
          in: query
          schema: { type: integer }
          description: Raw token units — returns notionalUsd + proceedsLamports
        - name: exitBps
          in: query
          schema: { type: integer }
          description: "1-10000 portion when amount is set (default 10000)"
      responses:
        "200":
          description: Quote (+ optional proceeds estimate)
        "429":
          description: Market data rate limited
        "503":
          description: Market data not configured
  /v1/tokens/{mint}/kline:
    get:
      tags: [Token · Market]
      summary: Get OHLCV price candles (chart)
      description: |
        Returns candlestick/OHLCV data for charting. Choose `resolution` and optional `from`/`to` (unix seconds). Also returns `lastClose` for a quick last price.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: mint
          in: path
          required: true
          schema: { type: string }
          description: Token mint address (Solana)
        - name: resolution
          in: query
          schema: { type: string, enum: [30s, 1m, 5m, 15m, 1h, 4h, 1d], default: 1m }
        - name: from
          in: query
          schema: { type: integer }
          description: Unix seconds
        - name: to
          in: query
          schema: { type: integer }
          description: Unix seconds
      responses:
        "200":
          description: Candles + lastClose
  /v1/tokens/{mint}/analyze:
    get:
      tags: [Token · Market]
      summary: Quick 1h price move analysis
      description: |
        Combines spot quote with roughly the last hour of kline data: open/close, % change, and volume. Use for a fast “is it pumping/dumping?” check.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: mint
          in: path
          required: true
          schema: { type: string }
          description: Token mint address (Solana)
      responses:
        "200":
          description: Analyze summary
  /v1/tokens/{mint}/info:
    get:
      tags: [Token · Research]
      summary: Token profile + market cap
      description: |
        Full token profile: name/symbol/decimals, liquidity, holders, supply, price, launchpad fields, and derived `marketCapUsd`.
        Also attaches DEX `pairs` / `bestPair` / `media` when available. Metadata fields are untrusted.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: mint
          in: path
          required: true
          schema: { type: string }
          description: Token mint address (Solana)
      responses:
        "200":
          description: Token info (metadata fields are untrusted)
  /v1/tokens/{mint}/security:
    get:
      tags: [Token · Research]
      summary: Token security / rug-risk metrics
      description: |
        Security audit-style fields (honeypot flags, renounce/mint/freeze status, and related risk signals). Use before buying to screen obvious rug/honeypot risks.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: mint
          in: path
          required: true
          schema: { type: string }
          description: Token mint address (Solana)
      responses:
        "200":
          description: Security payload
  /v1/tokens/{mint}/pool:
    get:
      tags: [Token · Research]
      summary: Liquidity pool details
      description: |
        Returns the main liquidity pool info for the token (pool address, reserves/liquidity context). Use when you need pool-level detail beyond the summary in `/info`.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: mint
          in: path
          required: true
          schema: { type: string }
          description: Token mint address (Solana)
      responses:
        "200":
          description: Pool payload
  /v1/tokens/{mint}/holders:
    get:
      tags: [Token · Research]
      summary: List top token holders
      description: |
        Paged/sorted list of top holders. Filter with `tag` (e.g. `dev`, `sniper`, `bundler`, `smart_degen`). Use to inspect who owns the supply.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: mint
          in: path
          required: true
          schema: { type: string }
          description: Token mint address (Solana)
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: orderBy
          in: query
          schema: { type: string, default: amount_percentage }
        - name: direction
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
        - name: tag
          in: query
          schema: { type: string, description: "smart_degen|renowned|dev|sniper|bundler|…" }
      responses:
        "200":
          description: Holders list
  /v1/tokens/{mint}/traders:
    get:
      tags: [Token · Research]
      summary: List top traders on this token
      description: |
        Top traders by activity/profit for the token. Use to see who is actively buying/selling, not just who currently holds.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: mint
          in: path
          required: true
          schema: { type: string }
          description: Token mint address (Solana)
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: orderBy
          in: query
          schema: { type: string, default: amount_percentage }
        - name: direction
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
        - name: tag
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Traders list
  /v1/tokens/{mint}/research:
    get:
      tags: [Token · Research]
      summary: One-shot due-diligence bundle
      description: |
        Convenience bundle: token info + security + pool, plus DEX `pairs` / `bestPair` when available. Prefer `/detail` for the fullest merged payload.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: mint
          in: path
          required: true
          schema: { type: string }
          description: Token mint address (Solana)
      responses:
        "200":
          description: Research summary
  /v1/tokens/{mint}/holder-analysis:
    get:
      tags: [Token · Research]
      summary: Holder chip-structure analysis + rating
      description: |
        Deep holder analysis: concentration (top10/20), burn/DEX vs normal supply, airdrop & risk wallets (bundler/sniper/rat), related/sybil funding clusters, smart money/KOL/whale signals, Dev holding, buying-power buckets, and an overall buy-rating (`normal` / `light` / `caution` / `not_recommended`). Use to judge if holder structure looks healthy before entering.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: mint
          in: path
          required: true
          schema: { type: string }
          description: Token mint address (Solana)
      responses:
        "200":
          description: Concentration, risk, related wallets, rating
  /v1/tokens/{mint}/detail:
    get:
      tags: [Token · Research]
      summary: Full token detail (research + pairs merged)
      description: |
        Combined token detail: research profile/security/pool plus all DEX pairs, best pair, media, and paid listing orders.
        Use this for a complete token page in one call.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: mint
          in: path
          required: true
          schema: { type: string }
          description: Token mint address (Solana)
      responses:
        "200":
          description: Merged token detail
  /v1/tokens/{mint}/pairs:
    get:
      tags: [Token · Research]
      summary: All DEX pairs for a token
      description: |
        Lists liquidity pairs for the mint (price, volume, liquidity, txns). Includes `bestPair` by liquidity.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: mint
          in: path
          required: true
          schema: { type: string }
          description: Token mint address (Solana)
        - name: chainId
          in: query
          schema: { type: string, default: solana }
      responses:
        "200":
          description: Pairs + bestPair
  /v1/tokens/{mint}/orders:
    get:
      tags: [Token · Research]
      summary: Token listing / boost orders
      description: |
        Paid listing or boost orders associated with the token (if any).
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: mint
          in: path
          required: true
          schema: { type: string }
          description: Token mint address (Solana)
        - name: chainId
          in: query
          schema: { type: string, default: solana }
      responses:
        "200":
          description: Orders payload
  /v1/discover/profiles/latest:
    get:
      tags: [Discover]
      summary: Latest token profiles
      description: |
        Newest curated Solana token profiles. Each item includes `tokenMetadata` from token detail when available.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
      responses:
        "200":
          description: Profile list
  /v1/discover/profiles/recent:
    get:
      tags: [Discover]
      summary: Recently updated token profiles
      description: |
        Token profiles that were recently edited/updated.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
      responses:
        "200":
          description: Profile list
  /v1/discover/takeovers/latest:
    get:
      tags: [Discover]
      summary: Latest community takeovers
      description: |
        Tokens with recent community takeover claims.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
      responses:
        "200":
          description: Takeover list
  /v1/discover/ads/latest:
    get:
      tags: [Discover]
      summary: Latest ads
      description: |
        Latest Solana promoted token ads. Each ad includes `tokenMetadata` for its `tokenAddress`.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
      responses:
        "200":
          description: Ads list
  /v1/discover/boosts/latest:
    get:
      tags: [Discover]
      summary: Latest boosted tokens
      description: |
        Recently boosted tokens (amount / totalAmount).
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
      responses:
        "200":
          description: Boost list
  /v1/discover/boosts/top:
    get:
      tags: [Discover]
      summary: Top actively boosted tokens
      description: |
        Tokens with the most active boosts right now.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
      responses:
        "200":
          description: Boost list
  /v1/discover/search:
    get:
      tags: [Discover]
      summary: Search pairs by name/symbol/address
      description: |
        Full-text style pair search. Pass `q` (token name, symbol, or address).
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: q
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Matching pairs
  /v1/discover/metas/trending:
    get:
      tags: [Discover]
      summary: Trending metas / narratives
      description: |
        Trending market metas (narratives) with marketCap, liquidity, volume, and tokenCount.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
      responses:
        "200":
          description: Meta list
  /v1/discover/metas/{slug}:
    get:
      tags: [Discover]
      summary: Meta detail + pairs
      description: |
        One meta/narrative by slug, including related pairs.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Meta with pairs
  /v1/discover/pairs/{chainId}/{pairId}:
    get:
      tags: [Discover]
      summary: Get one pair by chain + pair address
      description: |
        Returns pair market data (price, volume, liquidity, txns) for a specific pool.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: chainId
          in: path
          required: true
          schema: { type: string, example: solana }
        - name: pairId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Pair payload
  /v1/stream/profiles/latest:
    get:
      tags: [Live · Stream]
      summary: Live stream — latest profiles (WebSocket)
      description: |
        **WebSocket endpoint** — connect with `ws://` / `wss://` (Swagger "Try it out" uses plain HTTP and returns a JSON guide + snapshot).
        No `cluster` query required. Live updates require a WebSocket client.
      responses:
        "101":
          description: Switching Protocols (WebSocket)
        "200":
          description: JSON guide + optional snapshot when called without WebSocket Upgrade
  /v1/stream/profiles/recent:
    get:
      tags: [Live · Stream]
      summary: Live stream — recent profile updates (WebSocket)
      description: |
        **WebSocket** — use `ws://` / `wss://`. Plain HTTP returns a JSON guide + snapshot.
      responses:
        "101":
          description: Switching Protocols (WebSocket)
        "200":
          description: JSON guide + optional snapshot when called without WebSocket Upgrade
  /v1/stream/takeovers/latest:
    get:
      tags: [Live · Stream]
      summary: Live stream — community takeovers (WebSocket)
      description: |
        **WebSocket** — use `ws://` / `wss://`. Plain HTTP returns a JSON guide + snapshot.
      responses:
        "101":
          description: Switching Protocols (WebSocket)
        "200":
          description: JSON guide + optional snapshot when called without WebSocket Upgrade
  /v1/stream/ads/latest:
    get:
      tags: [Live · Stream]
      summary: Live stream — ads (WebSocket)
      description: |
        **WebSocket** — use `ws://` / `wss://`. Plain HTTP returns a JSON guide + snapshot.
      responses:
        "101":
          description: Switching Protocols (WebSocket)
        "200":
          description: JSON guide + optional snapshot when called without WebSocket Upgrade
  /v1/stream/boosts/latest:
    get:
      tags: [Live · Stream]
      summary: Live stream — latest boosts (WebSocket)
      description: |
        **WebSocket** — use `ws://` / `wss://`. Plain HTTP returns a JSON guide + snapshot.
      responses:
        "101":
          description: Switching Protocols (WebSocket)
        "200":
          description: JSON guide + optional snapshot when called without WebSocket Upgrade
  /v1/stream/boosts/top:
    get:
      tags: [Live · Stream]
      summary: Live stream — top boosts (WebSocket)
      description: |
        **WebSocket** — use `ws://` / `wss://`. Plain HTTP returns a JSON guide + snapshot.
      responses:
        "101":
          description: Switching Protocols (WebSocket)
        "200":
          description: JSON guide + optional snapshot when called without WebSocket Upgrade
  /v1/wallets/{walletAddress}/kind:
    get:
      tags: [Wallet · Analytics]
      summary: Classify wallet as EOA or PDA
      description: |
        Returns `walletKind` for an address:
        - `eoa` — individual / personal wallet
        - `pda` — vault wallet (program-derived)
        Pass optional `walletKind` to force; otherwise known vault pubkeys auto-resolve to `pda`.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - $ref: "#/components/parameters/WalletAddressPath"
        - $ref: "#/components/parameters/WalletKindQuery"
      responses:
        "200":
          description: "Classification payload (`walletKind`, `walletKindLabel`, `walletKindMeaning`)"
  /v1/wallets/{walletAddress}/holdings:
    get:
      tags: [Wallet · Analytics]
      summary: Wallet token holdings with P&L
      description: |
        Lists open (and optionally closed) token positions for a wallet with cost basis, USD value, realized/unrealized P&L. Requires market signing key. Includes `walletKind` (`eoa` individual / `pda` vault). Use for portfolio screens and copy-trade research.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - $ref: "#/components/parameters/WalletKindQuery"
        - $ref: "#/components/parameters/WalletAddressPath"
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 50 }
        - name: orderBy
          in: query
          schema: { type: string }
        - name: direction
          in: query
          schema: { type: string, enum: [asc, desc] }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Holdings payload
        "503":
          description: Market signing key not configured
...
  /v1/vaults:
    get:
      tags: [Vaults]
      summary: List indexed vaults
      description: |
        Paginated list of vaults known to the indexer for the cluster. Entry point for vault discovery/browse UI.
        Each item includes `vaultType` / `vaultTypeLabel` / `vaultTypeMeaning` (default `pooled`).
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - $ref: "#/components/parameters/VaultTypeQuery"
      responses:
        "200":
          description: Paginated vaults
...